requirements.txt
streamlit
pypdf2
langchain
langchain-community
python-dotenv
faiss-cpu
openai